Update basic control for Course

+++ src/main/java/com/therealdanvega/repository/CourseRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.therealdanvega.domain.Course;
 
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 
 	List<Course> findAllByOrderByCourseCodeAsc();
+
 }
